added prompt generation method working

+++ proj/src/modules/game/game.c
@@ -3,7 +3,7 @@
 static player_drawer_t *player_drawer;
 static canvas_t *canvas;
 static guess_word_t *guess;
-char *prompt;
+static char prompt[15];
 
 int (setup_game)(bool isTransmitter) {
   player_drawer = create_player_drawer(isTransmitter ? SELF_PLAYER : OTHER_PLAYER);
@@ -13,9 +13,11 @@ int (setup_game)(bool isTransmitter) {
   }
   canvas = canvas_init();
   guess = create_guess_word();
-  //if (prompt_generate(prompt) != OK) return EXIT_FAILURE;
+  printf("guess size: %d\n", guess->pointer);
+  if (prompt_generate(prompt) != OK) return EXIT_FAILURE;
   if (canvas == NULL || guess == NULL) {
     destroy_player_drawer(player_drawer);
+    destroy_guess_word(guess);
     return EXIT_FAILURE;
   }
   return EXIT_SUCCESS;
@@ -25,7 +27,6 @@ void (destroy_game)() {
   destroy_player_drawer(player_drawer);
   canvas_destroy(canvas);
   destroy_guess_word(guess);
-  //prompt_destroy(prompt);
   vg_clear_buffers();
 }
 
@@ -34,12 +35,18 @@ extern uint8_t scancode;
 int (game_process_keyboard)(){
   if (player_drawer_get_state(player_drawer) == SELF_PLAYER) return EXIT_SUCCESS;
   if (keyboard_return_value) return EXIT_SUCCESS;
+  if (scancode == 0) return EXIT_SUCCESS;
 
+  //printf("trying to process kbd\n");
   if (scancode == MAKE_BACKSPACE){
+    printf("apagando\n");
+    printf("old size: %d", guess->pointer);
     delete_character(guess);
+    printf("new size: %d\n", guess->pointer);
   }
   else if (scancode == MAKE_ENTER){
     bool right;
+    printf("validando\n");
     validate_guess_word(prompt, guess, &right);
     printf("%d \n", right);
     reset_guess_word(guess);
@@ -54,10 +61,14 @@ int (game_process_keyboard)(){
       uint8_t caracter = 0;
       if (translate_scancode(scancode, &caracter)) return EXIT_FAILURE;
       if (caracter != 0){
+        printf("writing\n");
+        printf("old size: %d ", guess->pointer);
         if (write_character(guess, caracter)) return EXIT_FAILURE;
+        printf("new size: %d\n", guess->pointer);
       }
     }
   }
+  //printf("processed kbd\n");
   return EXIT_SUCCESS;
 }
